Add support for handling 'sentinel booleans'
Some C functions with out arguments additionally return a boolean that
indicates whether the out arguments have been set.  In Perl, we usually want to
map that to a function that returns the out arguments if the boolean is true,
and an empty list otherwise.

Add support for this kind of mangling to Glib::Object::Introspection->setup.