Use a different approach to invoking fallback vfuncs
Instead of having one Perl fallback vfunc in the common ancestor package and
let it try to determine which class' vfunc to actually invoke, put a separate
Perl fallback vfunc into each parent package with implicit knowledge of which
class' vfunc to invoke.

Thanks for Kevin Ryde for suggesting this approach.