Do not assume filenames are UTF8-encoded
Instead of automatically converting filenames to and from UTF8-encoded Perl
strings, be agnostic about the encoding.  Before passing a filename on to
something which expects UTF8 (like widgets), users now have to ensure manually
that it is UTF8-encoded (via Glib::filename_to_unicode, for example).  Thus,
this change might be an API break.