Fix a marshalling bug with Perl code and array args
When invoking Perl code with array args, use the correct type for the length
arg.  Previously, we used the type of the array arg to convert the length arg
to a Perl value.  That is, in the most common case, we treated an integer as a
pointer.  This works OK on little-endian systems, but most likely breaks on
big-endian ones.