Copy item memory in flat arrays when we are given ownership
The item memory in flat arrays is usually allocated en-bloc, so we
cannot simply assume ownership of each individual array element, as that
would lead to deallocation attempts of sub-blocks of the big
allocation.  So instead we copy the individual items and free the array
memory en-bloc afterwards.

This happened for poppler_annot_text_markup_get_quadrilaterals.

https://rt.cpan.org/Public/Bug/Display.html?id=130280
https://gitlab.gnome.org/GNOME/perl-glib-object-introspection/issues/1